feat(ui): tool cost column on Top tools + MCP servers card
TopToolsCard adds a sortable 'cost' column from /api/tool-costs, defaults sort to descending cost (highest burners first), and shows a small error dot when the tool had failures. New McpServersCard renders a strip of MCP servers with their attributed cost and call count, rolled up from the same endpoint.

+++ crates/token-dashboard-core/src/tool_costs.rs
@@ -75,8 +75,8 @@ pub fn report<P: AsRef<Path>>(db: P, days: u32) -> rusqlite::Result<ToolCostRepo
             r.get::<_, i64>(1)? as u64,
             r.get::<_, i64>(2)? as u64,
             r.get::<_, Option<String>>(3)?,
-            r.get::<_, i64>(4)? as i64,
-            r.get::<_, i64>(5)? as i64,
+            r.get::<_, i64>(4)?,
+            r.get::<_, i64>(5)?,
         ))
     })?;
